Support greedy outputs
If 'is_greedy' is set on an addressee in the transaction json the amount
for that addressee is ignored and instead the remaining coins from the
transaction inputs after deducting the fee and the other
addresees/outputs are sent to the greedy addresee.

Only one greedy addressee can be specified per asset.

If a greedy addressee is specified it implies there will be no change
output.