Reimplement transaction fetching to use a persistent cache (9)
Store the latest seen block number in the local cache. When logging in,
remove mempool txs and any txs that could have been reorged leading up
to the last seen block.

Typically this means that over time, the wallet will make fewer and fewer
requests for txs as a proportion of the wallet contents. When wallet txs
are outstanding in the mempool there will be requests reasonably often,
but these requests need to return fewer items than the previous cache
impl.